Fixed the algorithm used to make the player face the mouse direction.
Fixed the GetForwardVector function to conform with the player/mouse direction

Separated scaling, rotation and translation into three separate matrices and made sure that the world matrix was multiplied in the correct order.

NOTE: We should consider exporting/importing the three matrices separately instead of the worldMatrix as it is.